auth added for all endpoints

+++ Tweets/views.py
@@ -1,9 +1,8 @@
-from django.shortcuts import render
 from datetime import datetime
-from django.shortcuts import render
 from TUsers.models import TUser
+from TUsers.views import is_autherized
 from Tweets.models import TCtweets
-from django.http import HttpResponse,QueryDict
+from django.http import HttpResponse
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
@@ -15,7 +14,7 @@ import logging
 logger = logging.getLogger(__name__)
 logging.basicConfig(filename='debug.log',level=logging.DEBUG)
 
-def get_user_obj(user):
+def get_user_obj(user=None):
     #Returns User Objects corresponding to the username
     return TUser.objects.get(username=user)
 
@@ -28,28 +27,32 @@ def CreateTweet(request):
     tweet_text: (mandatory) <str> Tweet
     Output: TCtweet Object of the created tweet
     '''
-
     if request.method == "POST":
         username = request.query_params.get('username')
         text = request.query_params.get('tweet_text')
-        validate = TCValidator(request.query_params,request.FILES)
-        if not validate.is_valid():
-            error = {'Error_code': status.HTTP_400_BAD_REQUEST,
-                        'Error_Message': "Invalid username or tweet_text"}
-            logger.error(error)                    
-            return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)
+        if is_autherized(request,username):
+            validate = TCValidator(request.query_params,request.FILES)
+            if not validate.is_valid():
+                error = {'Error_code': status.HTTP_400_BAD_REQUEST,
+                            'Error_Message': "Invalid username or tweet_text"}
+                logger.error(error)                    
+                return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)
 
-        try:
-            user = get_user_obj(username)
-            new_tweet = TCtweets(username=user,tweet_text=text)
-            new_tweet.save()
-            serializer = TCtweetsSerializer(new_tweet)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        except:
+            try:
+                user = get_user_obj(username)
+                new_tweet = TCtweets(username=user,tweet_text=text)
+                new_tweet.save()
+                serializer = TCtweetsSerializer(new_tweet)
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            except Exception as e:
+                error = {'Error_code': status.HTTP_400_BAD_REQUEST,
+                            'Error_Message': "User Does not exist"}
+                logger.error(e)                    
+                return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)
+        else:
             error = {'Error_code': status.HTTP_400_BAD_REQUEST,
-                        'Error_Message': "User Does not exist"}
-            logger.error(error)                    
-            return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)
+                                'Error_Message': "Authentication failed. Please login"}
+            return Response(error, status=status.HTTP_400_BAD_REQUEST)        
 
 @api_view(['GET'])
 def Timeline(request,username):
@@ -58,31 +61,34 @@ def Timeline(request,username):
     Input: page (mandatory) <int>  
     Output: Tweet object with all the tweets in the page
     '''
-
-    try:
-        page_no = int(request.query_params.get('page',1))
-        userObj = get_user_obj(username)
-        if page_no < 1:
+    if is_autherized(request,username):
+        try:
+            page_no = int(request.query_params.get('page',1))
+            userObj = get_user_obj(username)
+            if page_no < 1:
+                error = {'Error_code': status.HTTP_400_BAD_REQUEST,
+                            'Error_Message': "Please pass an integer value (starting with 1) as Page Number"}
+                logger.error(error)                    
+                return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)
+            
+            tweets = TCtweets.objects.filter(username=userObj)
+            paginator = Paginator(tweets,10) #shows 10 tweets per page
+            page_num = paginator.get_page(page_no)
+            tweet_objs = page_num.object_list 
+            serializer = TCtweetsSerializer(tweet_objs,many=True)
+            return Response(serializer.data)
+        except Exception as e:
             error = {'Error_code': status.HTTP_400_BAD_REQUEST,
-                        'Error_Message': "Please pass an integer value (starting with 1) as Page Number"}
-            logger.warning(error)                    
-            return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)
-        
-        tweets = TCtweets.objects.filter(username=userObj)
-        logger.warning(tweets)
-        paginator = Paginator(tweets,10) #shows 10 tweets per page
-        page_num = paginator.get_page(page_no)
-        tweet_objs = page_num.object_list 
-        serializer = TCtweetsSerializer(tweet_objs,many=True)
-        return Response(serializer.data)
-    except Exception as e:
+                        'Error_Message': "No Tweets to show"}
+            logger.error(e)                    
+            return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST) 
+    else:
         error = {'Error_code': status.HTTP_400_BAD_REQUEST,
-                    'Error_Message': "No Tweets to show"}
-        logger.error(e)                    
-        return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST) 
+                            'Error_Message': "Authentication failed. Please login"}
+        return Response(error, status=status.HTTP_400_BAD_REQUEST)        
 
 @api_view(['DELETE'])
-def DeleteTweet(request,tweet_id):
+def DeleteTweet(request,tweet_id=None):
     '''
     Purpose: Deletes the tweet having the id in the URL
     Input: None  
@@ -90,17 +96,23 @@ def DeleteTweet(request,tweet_id):
     '''
     try:
         tweet = TCtweets.objects.get(id=tweet_id)
-        serializer = TCtweetsSerializer(tweet)
-        tweet.delete()
-        return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        username = tweet.username.username
+        if is_autherized(request,username):
+            serializer = TCtweetsSerializer(tweet)
+            tweet.delete()
+            return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        else:
+            error = {'Error_code': status.HTTP_400_BAD_REQUEST,
+                                'Error_Message': "Authentication failed. Please login"}
+            return Response(error, status=status.HTTP_400_BAD_REQUEST)       
     except Exception as e:
         error = {'Error_code': status.HTTP_400_BAD_REQUEST,
                     'Error_Message': "This tweet no longer exists"}
         logger.error(e)    
-        return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)  
+        return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)     
 
 @api_view(['GET'])
-def ShowTweet(request,tweet_id):
+def ShowTweet(request,tweet_id=None):
     '''
     Purpose: displays the tweet with the id in the URL and its replies
     Input: None  
@@ -108,10 +120,16 @@ def ShowTweet(request,tweet_id):
     '''
     try:
         tweet = TCtweets.objects.filter(id=tweet_id)
-        replies = TCtweets.objects.get(id=tweet_id).reply.all()
-        tweetNreply = tweet.union(replies)
-        serializer = TCtweetsSerializer(tweetNreply,many=True)
-        return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        username = tweet.username.username
+        if is_autherized(request,username):
+            replies = TCtweets.objects.get(id=tweet_id).reply.all()
+            tweetNreply = tweet.union(replies)
+            serializer = TCtweetsSerializer(tweetNreply,many=True)
+            return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        else:
+            error = {'Error_code': status.HTTP_400_BAD_REQUEST,
+                                'Error_Message': "Authentication failed. Please login"}
+            return Response(error, status=status.HTTP_400_BAD_REQUEST)
     except Exception as e:
         logger.error(e)
         error = {'Error_code': status.HTTP_400_BAD_REQUEST,
@@ -119,7 +137,7 @@ def ShowTweet(request,tweet_id):
         return Response(error, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['PUT'])
-def Reply(request,tweet_id):
+def Reply(request,tweet_id=None):
     '''
     Purpose: Reply to the tweet with the id in the URL
     Input: username (mandatory) <str> Account user
@@ -135,15 +153,19 @@ def Reply(request,tweet_id):
                         'Error_Message': "Invalid username or reply_text"}
             logger.error(error)                    
             return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)
-
-        user =get_user_obj(user)
-        reply_tweet = TCtweets(username=user,tweet_text=reply)
-        reply_tweet.save()
-        tweet = TCtweets.objects.get(id=tweet_id)
-        tweet.reply.add(reply_tweet)
-        tweet.save()
-        serializer = TCtweetsSerializer(reply_tweet)
-        return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        if is_autherized(request,user):
+            user =get_user_obj(user)
+            reply_tweet = TCtweets(username=user,tweet_text=reply)
+            reply_tweet.save()
+            tweet = TCtweets.objects.get(id=tweet_id)
+            tweet.reply.add(reply_tweet)
+            tweet.save()
+            serializer = TCtweetsSerializer(reply_tweet)
+            return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        else:
+            error = {'Error_code': status.HTTP_400_BAD_REQUEST,
+                                'Error_Message': "Authentication failed. Please login"}
+            return Response(error, status=status.HTTP_400_BAD_REQUEST)
     except Exception as e:
         error = {'Error_code': status.HTTP_400_BAD_REQUEST,
                         'Error_Message': "Error saving the reply"}
@@ -151,7 +173,7 @@ def Reply(request,tweet_id):
         return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST) 
 
 @api_view(['PUT'])
-def Retweet(request,tweet_id):
+def Retweet(request,tweet_id=None):
     '''
     Purpose: Retweet the tweet with the id in the URL
     Input: username (mandatory) <str> Account user
@@ -168,14 +190,19 @@ def Retweet(request,tweet_id):
             logger.error(error)                    
             return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)
 
-        tweet_text = TCtweets.objects.get(id=tweet_id).tweet_text
-        reply_tweet = TCtweets(username = get_user_obj(user),tweet_text=tweet_text,comment=comment)
-        reply_tweet.save()
-        tweet= TCtweets.objects.get(id=tweet_id)
-        tweet.retweet.add(get_user_obj(user))
-        tweet.save()
-        serializer = TCtweetsSerializer(reply_tweet)
-        return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        if is_autherized(request,user):
+            tweet_text = TCtweets.objects.get(id=tweet_id).tweet_text
+            reply_tweet = TCtweets(username = get_user_obj(user),tweet_text=tweet_text,comment=comment)
+            reply_tweet.save()
+            tweet= TCtweets.objects.get(id=tweet_id)
+            tweet.retweet.add(get_user_obj(user))
+            tweet.save()
+            serializer = TCtweetsSerializer(reply_tweet)
+            return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        else:
+            error = {'Error_code': status.HTTP_400_BAD_REQUEST,
+                                'Error_Message': "Authentication failed. Please login"}
+            return Response(error, status=status.HTTP_400_BAD_REQUEST)
     except Exception as e:
         error = {'Error_code': status.HTTP_400_BAD_REQUEST,
                         'Error_Message': "Error retweeting!"}
@@ -183,7 +210,7 @@ def Retweet(request,tweet_id):
         return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST) 
 
 @api_view(['PUT'])
-def Like(request,tweet_id):
+def Like(request,tweet_id=None):
     '''
     Purpose: Like the tweet with the id in the URL
     Input: username (mandatory) <str> Account user 
@@ -197,14 +224,28 @@ def Like(request,tweet_id):
                         'Error_Message': "Invalid username"}
             logger.error(error)                    
             return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST)
-
-        tweet = TCtweets.objects.get(id=tweet_id)
-        tweet.like.add(get_user_obj(user))
-        tweet.save()
-        serializer = TCtweetsSerializer(tweet)
-        return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+            
+        if is_autherized(request,user):
+            tweet = TCtweets.objects.get(id=tweet_id)
+            tweet.like.add(get_user_obj(user))
+            tweet.save()
+            serializer = TCtweetsSerializer(tweet)
+            return Response(serializer.data, status=status.HTTP_204_NO_CONTENT)
+        else:
+            error = {'Error_code': status.HTTP_400_BAD_REQUEST,
+                                'Error_Message': "Authentication failed. Please login"}
+            return Response(error, status=status.HTTP_400_BAD_REQUEST)
     except Exception as e:
         error = {'Error_code': status.HTTP_400_BAD_REQUEST,
                         'Error_Message': "Error liking the tweet!"}
         logger.error(e)                    
         return Response(json.dumps(error), status=status.HTTP_400_BAD_REQUEST) 
+
+@api_view(['GET'])
+def all_tweets(request):
+    '''
+    Debugging
+    '''
+    tweets = TCtweets.objects.all()
+    serializer = TCtweetsSerializer(tweets,many=True)
+    return Response(serializer.data,status=status.HTTP_200_OK)
